+++ pages/_app.tsx
@@ -9,7 +9,7 @@ import { useEffect } from 'react'
 import { Provider } from 'react-redux'
 import { PersistGate } from 'redux-persist/integration/react'
 
-import BasicLayout from '@/components/layout/BasicLayout'
+import BasicLayout from '@/components/navBar/BasicLayout'
 import { persistor, store } from '@/redux/store'
 import theme from '@/styles/theme'
